Fix bug of meta incompatibility between starrocks-1.18 and starrocks-1.19 (#823)
Before starrocks-1.19, the meta key stored in json is dorisDBVersion. In the starrocks-1.19, the meta key changed to starrocksVersion. The upgrading is not compatible between starrocks-1.18 and starrocks-1.19. As a result, the upgrading will be crash. In this pull request, I change the logic. By double reading meta according to the two meta key, it can guarantee the meta can be read rightly in any case.