Support to access `char` type in hive orc file (#1176)
Char type in orc is quite nasty, because of the padding problem. And what's worse is that it uses space padding instead of zero padding.

To handle the char type in the orc file, we:

disable conjunct of char column. Because conjunct like a >= hello might be failed because value could be "hello "
remove trailing space in a string when read from orc file.
remove trailing space in min/max value in orc stats.