Make TMalloc GC Thread and CurrentThread can be exited gracefully. (#1383)
The TCMalloc GC Thread is not been joined when exited. It may be referenced resource already been released.
The CurrentThread thread local variable is used to gather the memory statistics and is defined in ExecEnv.
Because the exit of ExecEnv is after the StorageEngine. So we can check the exit state of StorageEngine to exit
the CurrentThread as soon as possible.