Support cancel delete after failure (#2927)
If the delete statement corresponds to many partitions, the remaining task may still exist after failure in the client.
This PR allows users to kill all FE and BE tasks.
When users use the MySQL client, they can use Ctrl+C to cancel this operation.
You can also use the following command.
```
kill $query_id
```
It's just that the client may retry the delete operation three times. At this time, it may need to kill three times.
All tasks will be automatically erased if the client is directly closed or the delete execution times out.

The task_worker_pool of BE has multiple tasks with different types.
Refactoring the TaskWorkerPool is necessary to support operation in one batch to accomplish this work.