[Enhancement]:Optimize bitmap index writer (#5329)
In the current BitmapIndexWriterImpl code, add_value needs to call twice getSizeInBytes the monitor the change of the roaring map. The add_value function is called very frequently. Which has a great impact on the performance of compaction in the test.
This pr optimizes the calculation of roaring size in the bitmap Index.

The solution to this problem is to move the calculation of the roaring bitmap size from the add_value functionto thesize` function. Delay calculation.