[Enhancement] Make segment iterator scan full chunk in each time  (#5701)
In each `_do_get_next`, segment iterators will return a result chunk with all of its elements passing the predicates _filter. To achieve this, it will scan some(chunk_capacity - chunk current size) rows, and filters these rows that are unsatisfied with the predicates. Assume the chunk size is 4096 when the chunk current size reaches 4095. Segment iterator will only scan more rows in `_read` and `_filter` it, when there are a number of rows unsatisfied with the predicates, this will have degradation to the performance.

To solve this problem, I make the segment iterator `_read` scan full rows(4096 rows, chunk_capacity) each time, and _filter 4096 rows in batch, which improves a lot to the scan speed. However, this may lead to the resulting chunk containing more than 4096 rows finally. We will save these overflow rows in _overflow_read_chunk chunk and reload them in next do_get_next.

This pr makes the scan speed has an overall improvement, in my test case, the best improvement is around 30%