[Tools] Add tablet meta consistency check tool (#5999)
We have met the bug which will cause inconsistency between tablet meta and segment footer(issue#4515). This bug has been fixed, but we can't determine if there are any problems with historical data unless we try to read the data even if BE is upgraded.

That is why add tablet meta consistency check in meta-tool. This function can check if there is an inconsistent tablet on BE.