[Feature] Add NTILE window function (#6246)
The NTILE window function divides ordered rows in the partition into `num_buckets` ranked groups of as equal size as possible and returns the group id of each row starting from 1.

It can not been used with the windowing clause. And for the implementation, it uses `ROWS BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW`.

The size of buckets could be `num_partition_rows/num_buckets` (small bucket) or `num_partition_rows/num_buckets` (large bucket). The top `num_partition_rows%num_buckets` buckets are the large buckets.

**NOTE** that, NTILE need the boundary of partition to calculate its value. For these functions, we must wait util the partition finished. Therefore, it uses the old version of `_process_by_partition_for_unbounded_preceding_rows_frame`.

The NTILE window function is only supported by the pipeline engine. BE will report the NotSupported error, when non-pipeline engine executes NTILE in BE.