[Enhancement] rank window function optimization, add threshold of partition number(4) (#6411)
For high cardinality of partition columns, each partition has only a few rows of data which result in performance degradation. So we can downgrade to stream mode if number of partition reaches threshold (the value is set to 4096 based on test ). 

Degradation is easy to implementation, because `PartitionTopN` makes no guarantees about the order of the output data, the only purpose of `PartitionTopN` is to filter data in order to alleviate the overhead of the following exchange and full sort operators.

```


                                   ┌────► topn─────┐
                                   │               │
 (unordered)                       │               │                  (unordered)
 inputChunks ───────► partitioner ─┼────► topn ────┼─► gather ─────► outputChunks
                                   │               │
                                   │               │
                                   └────► topn ────┘
```