[Refactor] Add types directory (#6873)
Now types related definitions are in different directories. Some of them are in the runtime directory, some of them are in the storage directory, some of them are in the utils directory. It is not friendly for our developers to understand. So I'm going to move all these definition to one directory.

In this PR, I create an types directory and move runtime/date_value.* and runtime/timestamp_value.* into this directory. I will move other definition in the following PRs.