[Feature] Compression Context Pool (#6879)
* [Feature] Compression Context Pool

Previously, each time we perform an LZ4F or ZSTD compression, we need to create an LZ4F or ZSTD compression context. Each compression context's memory size is not small, they are all allocated from the central cache of tcmalloc, not the local thread cache. For tcmlloac, allocating memory from the central cache needs to acquire the spinlock first. So when many threads create compression context together, they will compete to get the spinlock of tcmalloc's central cache, which will lead to the degradation of total performance.

In addition to the problem above, the previous compression will first allocate a very large buffer(equal to max_compressed_len) for compression output, and then shrink it after compression. Allocating and freeing large blocks of memory is done through the central cache of tcmalloc, which may also lead to a lot of conflicts.

For the first problem, I introduce a compression context pool for LZ4F and ZSTD, before compression we will request a compression context from the pool, after compression, we will return the context to the pool, in this way we can reuse the same context many times.

For the second problem, I add a compression buffer in each compression context, this compression buffer is shared under the same compression context, the compression result will first be saved on the compression buffer, and then copy to the target compression body. In this way, we can avoid allocating a very large buffer and then shrink it.