BugFix: fix data lost after compaction using persistent index (#8002)
When PersistentIndex is turned on, we use phmap as l0 to save kv pairs.
However, the kv pair storage may be discontinuous because phmap is aligned, which causes the wrong data to be written during flush. So the subsequent compaction may fail.
We will use uint8_t[8] instead of uint64_t to avoid the alignment of phmap and save memory usage.