[BugFix] add read lock to RoutineLoadManager to avoid log out of order (#8295)
1. A user creates a new routine job from a client.
2. The request is handled by a thread, transforming the SQL to a RoutineLoadJob and then put it into a map of the RoutineLoadManager class in a function guarded by a write lock.
3. The RoutineLoadTaskScheduler thread happens to start a new loop just then by getting all the routine load jobs from the map in the RoutineLoadManager class. Since there's no read lock, it finds the newly added job and schedulers it. Then it writes a journal on a state change of this new job.
4. The SQL executor thread is a bit slower in handling the new job. In the end, it writes a journal on creating a new job.
From the follower's view, the journal of creating a job is behind the journal of changing its state. As a result, it failed to replay the former journal because of failure to change the state of a non-existence job.

The solution is simple, just add a read lock would be enough. Since creating a routine job is not a frequently operation, it won't be too much cost for the scheduler thread.