Fix be crash in ASAN mode (#8722)
If element of ArrayColumn is empty, we don't need to read data from data_column of ArrayColumn because data is empty.
So we don't add an empty read_range into element_read_ranges, which will cause inconsistency between the begin of element_read_range and element_ordinal of element_iterator. That's why be crash in ASAN mode.