Upgrade librdkafka from 1.7.0 to 1.8.2 (#9133)
On version 1.7.0, query_watermark call is not transaction-aware. If we use read-committed, we will get the MaxOffset not the Last Stable offset. FE recorded the Last Stable offset but got MaxOffset, FE try to consume Kafka because of the lag of offset. But the read-committed transaction level will not consume any data. FE will retrying again and again. This comes from the producer write large of message but not committed. libkafka 1.8.2 resolves this problem and we upgrade to it.

edenhill/librdkafka#3423