Fix UT CreateTableTest.testAbnormal (#9211)
This PR is to fix the unit test problem introduced by #9182. The main reason is that under different parameters, the error prompts are different. There are actually 2 repair methods, 1. Change the single test error report 2. Change the configuration, this PR chooses A simpler way to change the configuration.If you think the way to report the error is better, you can also use another way to make changes.