[Enhancement] Avoid VERSION_INCOMPLETE repair task got blocked (#9573)
We should create a REPLICA_MISSING type task only when there exists enough available BEs which
we can choose to clone data. If not, we should check if we can create a VERSION_INCOMPLETE task
so that the repair of the replica with an incomplete version won't be blocked. Therefore the version publish process
of load task won't be blocked either.