[BugFix] Fix load encrypted kms failed with error: KMS provider at xxx threw an IOException ... Failed to find any Kerberos tgt (#11063)
This PR #8820 change the kerberos UGI (UserGroupInformation) to a private mode to support multi kerberos users. But for loading data from a kms encrypted path, calling kms server will fail because it cannot find a kerberos token. To fix this bug, set a kerberos ugi to the global context, so that kms client can use the kerberos token to interact with kms server.