[Feature] Add a new image formatter for RBAC based new privilege framework (#11528)
Introduce a new privilege framework.
To avoid huge PR which is hard to review and easy to conflict, I'm splitting into several small PR, each implementing an individual aspect of the new privilege framework. This is the second one. For previous PRs see
[Feature] new privilege framework: basic authentication #11447
This PR implements a new format of image and uses this format to persist the AuthenticationManager class.