[Feature] New RBAC framework: PrivilegeManager persist as image (#11777)
Introduce a new privilege framework.
To avoid huge PR which is hard to review and easy to conflict, I'm splitting into several small PR, each implementing an individual aspect of the new privilege framework. This is the fifth one. For previous PRs see
#11447
#11528
#11593
#11732

This PR implements the save() and load() function of the PrivilegeManager class so that it can be persisted in image.