[Refactor] Change value of primitive type (#13705)
There are too many types there. It is difficult to understand and to tell the differences. Also, it is error-prone.
Most Primitive and LogicalType are the same, and there is no need to keep them.
So I plan to merge these two into one type.
Because in ColumnPB, StarRocks use the value of LogicalType, we can't change the value of LogicalType.
But PrimitiveType's value is not persisted, so that we can change it according to the LogicalType value.
In this PR, I change the PrimitiveType value to its corresponding LogicalType.
In the following PR, I will make LogicalType have all PrimitiveType's functionality.
At last, I will make using PrimitiveType = LogicalType
Then there will be only LogicalType.