[BugFix] Fix the bug of empty table after recover table (#13921)
Problem:
This Problem was introduced by commit cb58cd4. This commit use Collections.unmodifiableSet instead of the new HashSet to reduce the memory usage of FE. But the problem here is that Collections.unmodifiableSet returns the view of nameToTable.keySet(). Before using the TableNames in recycleDatabase, nameToTable had been dropped, and also the view of it would become empty, which caused this problem.

Solution:
We copy the nameToTable.keySet() only when dropping the database.