[Bugfix] using json as TxnFinishState serde (#15290)
Originally, TxnFinishState is serialized using protobuf(baidu jprotobuf) and using the existing field reason in TxnState, when upgrading from version before 2.4 to 2.5, TxnFinishState serialized as binary string will be treated as reason string in older not upgraded FE, and the serialization of reason string will replace illegal chars in the string, which will corrupt TxnFinishState. So serialization using binary format is dangerous, and this PR change serialization format from protobuf to json.