[BugFix] Remove the closed hiveclient from client pool (#17533)
Currently, if a hive meta client that could access hive metastore need to be put into the client pool and reused. But if the connection has been closed, it will be also put back in the client pool. If user's hive metastore service ip keep changing, reconnecting the client using the original ip will cause time out. So we need to remove the closed hiveclient from client pool.