[Enhancement] Support delete sql for primary key table with sort key (#22799)
Primary key table with sort key does not support partial update so far, so we need to provide all columns data when we doing data import. But the delete sql only provide primary key columns and the delete request will be reject because the request is consider as partial update request. This pr will support the delete request for primary key table with sort key.