[BugFix] Unable to build from source, fix cmake error and use correct dependencies specified by third-party, clean unused variable. (#23112)
TLDR; 
Users can't build Starrocks from the source.  And Starrocks is now using incorrect dependencies which does not follow the restriction.

### All the libraries mentioned below are now used in starrocks in a wrong way (not the restricted version), starrocks is searching on users local environment instead of using specified third-party, which may cause unknown behavior.

**Furthermore, even if we successfully pack the project for distribution, the several packages inside the libraries may come from the "packing machine", not the wanted versions of the library indicated in Starrocks' third-party restrictions.**

Mainly because the wrong build script and a little unused variable.

More **IMPORTATNT**, this PR enable Starrocks use correct and specified dependencies indicated by third-party file. 

Before this, many packages are used from system or python environment. 

In issue #23111 , I described the situation in detail, including how to locate the bug and a description of the solution.
Errors:
1. CMake Error: Could NOT find Snappy (missing: Snappy_LIB)
2. CMakeError: Count NOT find RapidJSONAlt
3. Could NOT find Lz4 (missing: LZ4_LIB LZ4_INCLUDE_DIR)
4. Not using zlib in third-party
5. Brotli is also using library in the environment 
6. Copy files missing
7. Declared but not used variable




If a user can build from the source, then these dependencies are installed manually by apt on Ubuntu or yum on CentOS, which is strongly not recommended.