[Feature] Support PERCENT_RANK window function  (#22788)
Fixes #13021 #19632

The PERCENT_RANK() function calculates the relative rank of a row within
a result set as a percentage. It returns the percentage of partition
values less than the value in the current row, excluding the highest
value. The return values range from 0 to 1. This function is useful for
percentile calculations and analyzing data distribution.

The PERCENT_RANK() function is calculated using the following formula,
where rank represents the row rank and rows represents the number of
partition rows: