[BugFix] Remove incorrect exception throws (#27597)
If we enable replicated storage, we will try to pick a replica as the
primary replica. We will record the last write status and skip the write
fail replica. However, if all replica failed during last write, we
should pick a replica as primary replica at random. But we throw a
exception right now.


Signed-off-by: zhangqiang <qiangzh95@gmail.com>