[BugFix] Fix memory leak bug caused by label cleaner (#28311)
Fixes #27975

`loadMgr.cancelResidualJob()` will cancel the UNKNOWN state insert load
jobs, and write editLog for every changed job. It should be executed
only on leader node, because non-leader node cannot write editLog and
the thread will be blocked forever once non-leader node calls
`submitLog` (This is unreasonable, it will be changed by another PR).
To fix this bug move the cancelResidualJob into txn cleaner thread,
which is executed only on leader node.

Besides, the `cancelResidualJob` is based on `getLabelStatus`, but if
there is no txn related to that label (If the target table is system
table, there will be no txn related to the label), `getLabelStatus` will
return `UNKNOWN` too, which will cause the job cancelled by mistake. So
I added a new function `getLabelTransactionState`, and only when the txn
exists and its status is `UNKNOWN`, the job can be cancelled.

Signed-off-by: gengjun-git <gengjun@starrocks.com>