[Refactor] Add a new config enable_drop_tablet_if_unfinished_txn when dropping tablet (#30847)
When dropping tablet, there is some tablet still running with txns. The FE balance may drop the tablet, but the tablet still has txns. There is a sequence where the tablet will still receive the txn.
```
1. Drop the tablet
2. Receive a new query and plan
3. Begin a transaction
```
The BeginTransaction is after the query plan, it will make the ingestion can not get the accurate tablet state. But the ingestion will fail, and throw an
```
no delete vector found tablet
```
This config is just a temporary mechanism before changing the txn mechanism in the FE.