[BugFix] Add a dummy cache handle to avoid dependency error if build without starcache. (#35391)
Why I'm doing:
As we depend on the starcache object handle in parquet file reader to support footer cache, which make the starcache be a strong dependency to starrocks and must be included. So, if we build the starrocks with WITH_STARCACHE=OFF, there will be "no such file" error happen.

What I'm doing:
We create a dummy cache handle in this PR to decouple starrocks and starcache if build without starcache. So, when building starrocks with WITH_STARCACHE=OFF, we use the dummy cache handle instead of starcache object handle, to avoid the dpendency building errors.

Signed-off-by: Gavin <yangguansuo@starrocks.com>