[BugFix] Fix plan and alter concurrent issue (#49355)
## Why I'm doing:
Alter operation will modify table meta under db lock, and when alter operation and query are concurrent, if query is not hold the lock, we will check `lastSchemaUpdateTime` to make sure there are no alter operation during create query plan. However, using `lastSchemaUpdateTime` is not enough, the following is one case:
1. start to create logic plan at time t1.
2. start to alter table at time t2
3. modify table meta at time t3, but alter is not finished totally
4. start to create physical plan at time t4
5. create plan finished at time t5
6. alter operation finished at t6
7. t1 < t2 < t3 < t4 < t5 < t6

Because the modify table meta is not an atomic operation and we only update `lastSchemaUpdateTime` after alter operation is finished. So the meta of logic plan and physical plan maybe inconsistent.

## What I'm doing:
1. Add a new state `UPDATING_META`. When we start to modify table meta, the table state will be set as `UPDATING_META`
2. We will check the table state after creating exec plan. If the table state is UPDATING_META, we consider this plan is invalid
3. Keep the first `lastSchemaUpdateTime` before release lock. Because alter operation maybe finished during analyze and creating exec plan.

Fixes https://github.com/StarRocks/StarRocksTest/issues/8244

Signed-off-by: sevev <qiangzh95@gmail.com>