[Doc] unlisted English (#50494)
We have a few docs that are only published in the Chinese docs. These are specific to Chinese cloud providers or HDFS use cases. This PR adds them to the sidebar nav. They will not show up in the English docs as there are stub files in the English doc source that have unlisted: true in the frontmatter.

Signed-off-by: DanRoscigno <dan@roscigno.com>